+++ src/main/java/view/container/aspects/designs/board/ChessDesign.java
@@ -23,7 +23,7 @@ public class ChessDesign extends BoardDesign
         this.checkeredBoard = true;
         final SVGGraphics2D g2d = this.boardStyle.setSVGRenderingValues();
         final float swRatio = 0.005f;
-        final float swThin = (float)Math.max(1, (int)(0.005f * this.boardStyle.placement().width + 0.5));
+        final float swThin = Math.max(1, (int)(0.005f * this.boardStyle.placement().width + 0.5));
         final float swThick = 1.0f * swThin;
         this.setStrokesAndColours(context, new Color(0, 0, 0), new Color(150, 75, 0), new Color(200, 150, 75), new Color(250, 221, 144), new Color(223, 178, 110), new Color(255, 240, 200), null, swThin, swThick);
         this.drawGround(g2d, context, true);
