+++ src/main/java/view/container/aspects/designs/board/Connect4Design.java
@@ -27,7 +27,7 @@ public class Connect4Design extends BoardDesign
     @Override
     public String createSVGImage(final Context context) {
         final SVGGraphics2D g2d = this.boardStyle.setSVGRenderingValues();
-        this.setStrokesAndColours(context, new Color(120, 190, 240), new Color(125, 75, 0), new Color(210, 230, 255), null, null, null, new Color(0, 0, 0), (float)Math.max(1, (int)(0.0025 * this.boardStyle.placement().width + 0.5)), (float)(int)(2.0 * Math.max(1, (int)(0.0025 * this.boardStyle.placement().width + 0.5))));
+        this.setStrokesAndColours(context, new Color(120, 190, 240), new Color(125, 75, 0), new Color(210, 230, 255), null, null, null, new Color(0, 0, 0), Math.max(1, (int)(0.0025 * this.boardStyle.placement().width + 0.5)), (int)(2.0 * Math.max(1, (int)(0.0025 * this.boardStyle.placement().width + 0.5))));
         this.drawConnect4Board(g2d);
         this.topology().vertices().clear();
         this.topology().edges().clear();
